requirements.txt
aiohttp==3.7.4.post0
aiohttp_cors==0.7.0
async_timeout==3.0.1
chardet==4.0.0
dateparser==1.0.0
elasticsearch==7.15.1
gspread==5.3.0
itemadapter==0.2.0
kafka_python==2.0.2
oauth2client==4.1.3
oss2==2.14.0
pandas==1.2.4
pdfminer==20191125
pdfminer.six==20200517
pdfplumber==0.5.28
pygsheets==2.0.5
PyMySQL==1.0.2
python_dateutil==2.8.2
requests==2.25.1
Scrapy==2.5.0
xpinyin==0.7.6
cryptography==36.0.2
pyOpenSSL==22.0.0
regex==2021.4.4